Merged PR 1254: 👷 add pipeline for running autoscaler unit tests
Adds an ADO pipeline for running unit tests against the CA fork.

We've been actively making changes that are AKS RP-specific in that fork and so we are trying to make the experience better.

Changes that make sense in the OSS version of CA will always still be pushed there as well.

Related work items: #1286