cherrypick: Merged PR 1346: Realign azure template labels with AKS changes
realign scale from zero labels with AKS changes.

The poolName tag on VMSS had its key change to be prefixed with managed by. This means we lost the ability to predict the value for the `agentPool` label added in AKS.

This also adds a set of new labels that will replace older non-prefixed ones