Merged PR 1360: Remove check for returning in-memory size when VMSS is in updating state
We've had this logic for a while now. If the VMSS is updating state, return the in-memory size for a nodegroup/scale-set.

This was trying to guard from a scenario like below:
1. Autoscaler scales up to 4 instances
2. We get hit by a cache refresh
3. ARM hasn't registered our new capacity goal so we get stale data and we end up updating the in-memory cache to a lower value.

We've recently guarded against this case by this PR: https://dev.azure.com/AzureContainerUpstream/Kubernetes/_git/autoscaler/commit/49290b03589b4f739e855ac1f8b0e31eba371ee0?refName=refs%2Fheads%2Fcluster-autoscaler-release-1.22

where we extend our in-memory TTL so we should never hit that scenario.

Because our cache TTL is now 5 minutes, I believe this "Updating" logic can lead to a bad scenario where we take very long to reconcile with the real IaaS capacity goal if you for example get hit by multiple scale ups and it takes VMSS too long to scale or if cloudprovider is updating the network profile. In this case, it's best to not have it.