cherrypick: Merged PR 1896: Fix bug related to startInstances()
Fix bug related to startInstances()
This [ICM](https://portal.microsofticm.com/imp/v3/incidents/details/333703551/home) tracks the issue where CA fails to start a deallocated instance. CA only starts the deallocated instances that too when we do a scaleUp. More details on deallocation scaling can be found [here](https://msazure.visualstudio.com/CloudNativeCompute/_wiki/wikis/CloudNativeCompute.wiki/202806/Deallocate-Mode).

Root Cause:
CA proactively sets the status of the instance in it's instanceCacche to "Running" before waiting for startInstanceResult. This is done so as to improve CA's speed. But if the starting fails, the instance would have been still marked as "Running" but it should have been "Runniing" and then CA never tries to start this instance again in future because of incorrect state in it's state. Either failed provisioning or failure during startUp - 2 possible scenarios to deal with. This problem is not failed because during createUpdate and delete because of timely List() calls made.

Code changes:
1. Fix the bug. Check the status of the vm by making a list call if there was a failure during startUp. VMSS doesn't return the list of vms failed in the error. Hence, we're making a list call here.
2. Replace "azure://" with azurePrefix const. We use this prefix string at a lot of places in the code. It is better to use a const rather than the string always.

Related work items: #1671